Moved tests for ContractEntityRepository to common
